+++ src/components/DatePicker/DateRangePicker.story.js
@@ -1,3 +1,4 @@
+/* eslint-disable react/jsx-no-bind */
 import React from 'react';
 import styled from 'styled-components';
 import moment from 'moment';
@@ -80,13 +81,68 @@ stories
       />
     </Container>
   )))
-  .add('With past dates selectable', (() => (
+  .add('Allow any dates', (() => (
     <Container>
       <DateRangePicker
         endDateInputId='end-date-id-0'
         focusedInput='endDate'
         initialStartDate={ moment() }
+        isFutureDateSelectable={ boolean('isFutureDateSelectable', true) }
         isPastDateSelectable={ boolean('isPastDateSelectable', true) }
+        onDatesChange={ action('onDatesChange') }
+        startDateInputId='start-date-id-0'
+      />
+    </Container>
+  )))
+  .add('Only allow future dates', (() => (
+    <Container>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        focusedInput='endDate'
+        initialStartDate={ moment() }
+        isFutureDateSelectable={ boolean('isFutureDateSelectable', true) }
+        isPastDateSelectable={ boolean('isPastDateSelectable', false) }
+        onDatesChange={ action('onDatesChange') }
+        startDateInputId='start-date-id-0'
+      />
+    </Container>
+  )))
+  .add('Only allow past dates', (() => (
+    <Container>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        focusedInput='endDate'
+        initialVisibleMonth={ () => moment().subtract(1, 'months') }
+        isFutureDateSelectable={ boolean('isFutureDateSelectable', false) }
+        isPastDateSelectable={ boolean('isPastDateSelectable', true) }
+        onDatesChange={ action('onDatesChange') }
+        startDateInputId='start-date-id-0'
+      />
+    </Container>
+  )))
+  .add('With custom range checker', (() => (
+    <Container>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        focusedInput='endDate'
+        initialStartDate={ moment() }
+        isOutsideRange={ action('isOutsideRange') }
+        onDatesChange={ action('onDatesChange') }
+        startDateInputId='start-date-id-0'
+      />
+    </Container>
+  )))
+  .add('With an initial starting month', (() => (
+    <Container>
+      <p>Adjust the left hand month using a function like:
+        <code>  () => moment().subtract(1, 'months')</code> with the
+        <code>  initialVisibleMonth  </code> prop.
+      </p>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        focusedInput='endDate'
+        initialVisibleMonth={ () => moment().subtract(1, 'months') }
+        isOutsideRange={ action('isOutsideRange') }
         onDatesChange={ action('onDatesChange') }
         startDateInputId='start-date-id-0'
       />
@@ -97,7 +153,7 @@ stories
       <DateRangePicker
         endDateInputId='end-date-id-0'
         focusedInput='startDate'
-        keepOpenOnDateSelect={ boolean('keepOpenOndateSelect', true) }
+        keepOpenOnDateSelect={ boolean('keepOpenOnDateSelect', true) }
         onDatesChange={ action('onDatesChange') }
         startDateInputId='start-date-id-0'
       />
@@ -196,7 +252,6 @@ stories
       />
     </Container>
   )))
-  /* eslint-disable react/jsx-no-bind */
   .add('With panelButtons as render props function returning an array', (() => (
     <Container>
       <DateRangePicker
@@ -227,7 +282,6 @@ stories
         startDateInputId='start-date-id-0'
       />
     </Container>
-    /* eslint-enable react/jsx-no-bind */
   )));
 
 const Container = styled.div`